Move premakeDefines into the tools folder
