+++ GLFWWindowManager/premake5.lua
@@ -1,4 +1,4 @@
-include "../premakeDefines.lua"
+include "../Tools/Premake5/premakeDefines.lua"
 
 project "GLFWWindowManager"
 	kind "StaticLib"
