+++ Tools/Premake5/premakeDefines.lua
